% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin history}
\format{
An object of class \code{data.frame} with 3264 rows and 7 columns.
}
\usage{
bitcoin
}
\value{
\item{date}{The date}

\item{open}{Open bitcoin value in US dollars in the respective date}

\item{close}{Close bitcoin value in US dollars in the respective date}

\item{variation}{Percentage variation from open to close value}

\item{y}{Variation positive (1) or negative (0)}

\item{sample}{If the respetive date is part of the training, validation or test sample}

\item{x}{The past variation (positive or negative) on the prior 30 days}
}
\description{
Data set containing the bitcoin value history from April 30th 2013 to April 6th 2022.
}
\details{
The data is divided into training, validation and test samples.
}
\references{
Diego Marcondes. A data-driven systematic, consistent and feasible approach to Model Selection. \emph{PhD thesis}. Universidade de São Paulo, 2022
}
\keyword{datasets}
