% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_models.R
\name{error_models}
\alias{error_models}
\title{Predict output from multiple models}
\usage{
error_models(sample, predict, names, simulate = F, init_value = 1000)
}
\arguments{
\item{sample}{A data frame with the sample data. Should have collumns named x and y representing the input and output strings, and collumns variation and date containing the daily variation anda date for simulation.}

\item{predict}{A list of functions which, given an input string, returns a predicted output.}

\item{names}{A vector containing the names of the models which generated th predict functions.}
}
\value{
A list containing the sample data frame with new collumns for prdiction and simulation, the error of each model, a plot with the simularion and the data used in the plot.
}
\description{
Predict the output for each observation in a sample and calculate the error for multiple models.
}
\details{

}
