% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_LS.R
\name{search_LS}
\alias{search_LS}
\title{Search the Markov Boolean Learning Space}
\usage{
search_LS(
  train,
  val,
  k,
  max_leaves = 50,
  max_init,
  verbose = T,
  error = "simple"
)
}
\arguments{
\item{train}{A data frame with the training data. Should have columns named x and y representing the input and output strings.}

\item{val}{A data frame with the validation data. Should have columns named x and y representing the input and output strings.}

\item{k}{Maximum length of Markov Chain.}

\item{max_leaves}{Maximum number of leaves in a node.}

\item{max_init}{Maximum length of initial node to try.}

\item{verbose}{Logical. Whether to print algorithm progress.}

\item{error}{Wheter to consider the simple error or the weighted error}
}
\value{
A list containing the validation error of the estimated node, a table with the estimated hypothesis, a function to predict the output of a given input and the algorithm processing time.
}
\description{
Performs an stochastic search on the Markov Boolean Learning Space.
}
\details{
Performs the suboptimal algorithm discussed in the reference.
}
\examples{
mod <- search_LS(train = bitcoin[bitcoin$sample == "Train",],
val = bitcoin[bitcoin$sample == "Validation",],
k = 30,max_leaves = 8,max_init = log(8,2),verbose = FALSE)

}
\references{
Diego Marcondes. A data-driven systematic, consistent and feasible approach to Model Selection. \emph{PhD thesis}. Universidade de São Paulo, 2022
}
