% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERM_markov.R
\name{ERM_markov}
\alias{ERM_markov}
\title{ERM estimative of Markov model of length k}
\usage{
ERM_markov(train, k)
}
\arguments{
\item{train}{A data frame with the training data. Should have collumns named x and y representing the input and output strings.}

\item{k}{Length of Markov Chain.}
}
\value{
A list containing the table with the estimated hypothesis and a function to predict the output of a given input.
}
\description{
Estimate the hypothesis that minimizes the empirical error under the Markov model of length k.
}
\details{

}
